+++ example/android/app/src/main/kotlin/org/lichess/sound_effect_example/MainActivity.kt
@@ -2,4 +2,4 @@ package org.lichess.sound_effect_example
 
 import io.flutter.embedding.android.FlutterActivity
 
-class MainActivity: FlutterActivity()
+class MainActivity : FlutterActivity()