+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.0
+
+- Allow playing multiple sounds at the same time, with the `maxStreams`
+  parameter in `initialize`.
+- Run iOS platform side handlers in a background thread.
+
 ## 0.0.2
 
 - Make sure the sound is loaded when `load`'s Future is resolved on Android.